% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brg.R
\name{rgb_to_hex}
\alias{rgb_to_hex}
\title{RGB to Hex Color Converter}
\usage{
rgb_to_hex(r, g, b)
}
\arguments{
\item{r}{Red value (0-255)}

\item{g}{Green value (0-255)}

\item{b}{Blue value (0-255)}
}
\value{
A hex color string
}
\description{
RGB to Hex Color Converter
}
\keyword{internal}
