% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brg.R
\name{scale_color_brg}
\alias{scale_color_brg}
\title{BRG Color Scale}
\usage{
scale_color_brg(palette = "primary")
}
\arguments{
\item{palette}{The palette to use ("primary", "secondary", "tertiary", or "all")}
}
\value{
A ggplot2 color scale
}
\description{
Create a color scale using the BRG palette
}
