% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brg.R
\name{scale_fill_brg}
\alias{scale_fill_brg}
\title{BRG Fill Scale}
\usage{
scale_fill_brg(palette = "primary")
}
\arguments{
\item{palette}{The palette to use ("primary", "secondary", "tertiary", or "all")}
}
\value{
A ggplot2 fill scale
}
\description{
Create a fill scale using the BRG palette
}
