% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brg.R
\name{theme_brg}
\alias{theme_brg}
\title{BRG Custom Theme for ggplot2}
\usage{
theme_brg(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}
}
\value{
A ggplot2 theme object
}
\description{
A custom ggplot2 theme using the BRG color palette
}
