% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_brg.R
\docType{data}
\name{theme_brg_colors}
\alias{theme_brg_colors}
\title{BRG Theme Color Palettes}
\format{
An object of class \code{list} of length 3.
}
\usage{
theme_brg_colors
}
\description{
BRG Theme Color Palettes
}
\keyword{datasets}
